requirements.txt
flask
python-dotenv
langchain
langchain-community
pinecone
google-generativeai
python-docx
pypdf
docx2txt